Fix warnings, organize packages, add skipped tests
