+++ test/core/data/repo/activities_repo_test.dart
@@ -1,0 +1,2 @@
+void main() {
+}
