+++ test/core/data/repo/contacts_repo_test.dart
@@ -1,0 +1,2 @@
+void main() {
+}
