+++ test/core/data/repo/life_events_repo_test.dart
@@ -1,0 +1,2 @@
+void main() {
+}
