+++ test/core/data/repo/photos_repo_test.dart
@@ -1,0 +1,2 @@
+void main() {
+}
