+++ test/core/networking/binary_result_test.dart
@@ -1,0 +1,2 @@
+void main() {
+}
